+++ Authentication - Jason/src/components/Signup.tsx
@@ -1,63 +1,15 @@
-import { Link, useNavigate } from "react-router-dom";
-import { UserAuth } from "../context/AuthContext";
 import "./AccountCreation.css";
 import { useState } from "react";
+import SignupForm from "./SignupForm";
+import SignupSuccess from "./SignupSuccess";
 
 const Signup = () => {
-  const [loading, setLoading] = useState(false);
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [error, setError] = useState("");
-
-  const navigate = useNavigate();
-
-  const { signUpNewUser } = UserAuth();
-
-  const handleSignUp = async (e: any) => {
-    e.preventDefault();
-
-    setLoading(true);
-    try {
-      const result = await signUpNewUser(email, password);
-
-      if (result.success) {
-        navigate("/dashboard");
-        setLoading(false);
-      } else {
-        console.log(result.data);
-        throw new Error(result.data.message);
-      }
-    } catch (error) {
-      setError(`${error}`);
-      setLoading(false);
-    }
-  };
-
+  const [signupSuccess, setSignupSuccess] = useState(false);
+  // Displays message telling user to sign in with their email on successful sign up
   return (
     <div className="home-container">
       <div className="content-center-card">
-        <form className="account-form" onSubmit={(e) => handleSignUp(e)}>
-          <h2>Sign Up for MAGE Platform!</h2>
-          <p>
-            Have an account? Login <Link to="/signin">here</Link>
-          </p>
-          <div className="signup-input-container">
-            <input
-              onChange={(e) => setEmail(e.target.value)}
-              placeholder="Email"
-              type="email"
-            />
-            <input
-              onChange={(e) => setPassword(e.target.value)}
-              placeholder="Password"
-              type="password"
-            />
-            <button className="link-button" type="submit" disabled={loading}>
-              Sign Up
-            </button>
-          </div>
-        </form>
-        {error && <p>{error}</p>}
+        {signupSuccess ? <SignupSuccess/> : <SignupForm setSignupSuccess={setSignupSuccess}/>}
       </div>
     </div>
   );
